Move field value validation to the serializer.ts
