+++ dev/src/document.ts
@@ -19,22 +19,15 @@ const deepEqual = require('deep-equal');
 import * as assert from 'assert';
 
 import {google} from '../protos/firestore_proto_api';
-import {DeleteTransform, FieldTransform} from './field-value';
-import {GeoPoint} from './geo-point';
+import {FieldTransform} from './field-value';
 import {FieldPath, validateFieldPath} from './path';
 import {DocumentReference} from './reference';
 import {isPlainObject, Serializer} from './serializer';
 import {Timestamp} from './timestamp';
-import {ApiMapValue, DocumentData, UpdateMap, ValidationOptions} from './types';
+import {ApiMapValue, DocumentData, UpdateMap} from './types';
 import {isEmpty, isObject} from './util';
-import {customObjectMessage, invalidArgumentMessage} from './validate';
 
 import api = google.firestore.v1beta1;
-
-/*!
- * The maximum depth of a Firestore object.
- */
-const MAX_DEPTH = 20;
 
 /**
  * Returns a builder for DocumentSnapshot and QueryDocumentSnapshot instances.
@@ -1015,108 +1008,4 @@ export class Precondition {
   get isEmpty(): boolean {
     return this._exists === undefined && !this._lastUpdateTime;
   }
-}
-
-/**
- * Validates a JavaScript value for usage as a Firestore value.
- *
- * @private
- * @param arg The argument name or argument index (for varargs methods).
- * @param value JavaScript value to validate.
- * @param desc A description of the expected type.
- * @param path The field path to validate.
- * @param options Validation options
- * @param level The current depth of the traversal. This is used to decide
- * whether deletes are allowed in conjunction with `allowDeletes: root`.
- * @param inArray Whether we are inside an array.
- * @throws when the object is invalid.
- */
-export function validateUserInput(
-    arg: string|number, value: unknown, desc: string,
-    options: ValidationOptions, path?: FieldPath, level?: number,
-    inArray?: boolean): void {
-  if (path && path.size > MAX_DEPTH) {
-    throw new Error(
-        `${invalidArgumentMessage(arg, desc)} Input object is deeper than ${
-            MAX_DEPTH} levels or contains a cycle.`);
-  }
-
-  options = options || {};
-  level = level || 0;
-  inArray = inArray || false;
-
-  const fieldPathMessage = path ? ` (found in field ${path.toString()})` : '';
-
-  if (Array.isArray(value)) {
-    const arr = value as unknown[];
-    for (let i = 0; i < arr.length; ++i) {
-      validateUserInput(
-          arg, arr[i]!, desc, options,
-          path ? path.append(String(i)) : new FieldPath(String(i)), level + 1,
-          /* inArray= */ true);
-    }
-  } else if (isPlainObject(value)) {
-    const obj = value as object;
-    for (const prop in obj) {
-      if (obj.hasOwnProperty(prop)) {
-        validateUserInput(
-            arg, obj[prop]!, desc, options,
-            path ? path.append(new FieldPath(prop)) : new FieldPath(prop),
-            level + 1, inArray);
-      }
-    }
-  } else if (value === undefined) {
-    throw new Error(`${
-        invalidArgumentMessage(
-            arg, desc)} Cannot use "undefined" as a Firestore value${
-        fieldPathMessage}.`);
-  } else if (value instanceof DeleteTransform) {
-    if (inArray) {
-      throw new Error(`${invalidArgumentMessage(arg, desc)} ${
-          value.methodName}() cannot be used inside of an array${
-          fieldPathMessage}.`);
-    } else if (
-        (options.allowDeletes === 'root' && level !== 0) ||
-        options.allowDeletes === 'none') {
-      throw new Error(`${invalidArgumentMessage(arg, desc)} ${
-          value
-              .methodName}() must appear at the top-level and can only be used in update() or set() with {merge:true}${
-          fieldPathMessage}.`);
-    }
-  } else if (value instanceof FieldTransform) {
-    if (inArray) {
-      throw new Error(`${invalidArgumentMessage(arg, desc)} ${
-          value.methodName}() cannot be used inside of an array${
-          fieldPathMessage}.`);
-    } else if (!options.allowTransforms) {
-      throw new Error(`${invalidArgumentMessage(arg, desc)} ${
-          value.methodName}() can only be used in set(), create() or update()${
-          fieldPathMessage}.`);
-    }
-  } else if (value instanceof FieldPath) {
-    throw new Error(`${
-        invalidArgumentMessage(
-            arg,
-            desc)} Cannot use object of type "FieldPath" as a Firestore value${
-        fieldPathMessage}.`);
-  } else if (value instanceof DocumentReference) {
-    // Ok.
-  } else if (value instanceof GeoPoint) {
-    // Ok.
-  } else if (value instanceof Timestamp || value instanceof Date) {
-    // Ok.
-  } else if (value instanceof Buffer || value instanceof Uint8Array) {
-    // Ok.
-  } else if (value === null) {
-    // Ok.
-  } else if (typeof value === 'object') {
-    throw new Error(customObjectMessage(arg, value, path));
-  }
-}
-
-export function validateFieldValue(
-    arg: string|number, val: unknown, path?: FieldPath): void {
-  validateUserInput(
-      arg, val, 'Firestore value',
-      {allowDeletes: 'root', allowTransforms: true}, path);
 }
