+++ dev/src/serializer.ts
@@ -16,16 +16,29 @@
 
 import {google} from '../protos/firestore_proto_api';
 import {detectValueType} from './convert';
-import {FieldTransform} from './field-value';
+import {DeleteTransform, FieldTransform} from './field-value';
 import {GeoPoint} from './geo-point';
 import {DocumentReference, Firestore} from './index';
-import {ResourcePath} from './path';
+import {FieldPath, ResourcePath} from './path';
 import {Timestamp} from './timestamp';
 import {isEmpty, isObject} from './util';
 
 import api = google.firestore.v1beta1;
+import {ValidationOptions} from './types';
+import {customObjectMessage, invalidArgumentMessage} from './validate';
 
-/** An interface for Firestore types that can be serialized to Protobuf. */
+/**
+ * The maximum depth of a Firestore object.
+ *
+ * @private
+ */
+const MAX_DEPTH = 20;
+
+/**
+ * An interface for Firestore types that can be serialized to Protobuf.
+ *
+ * @private
+ */
 export interface Serializable {
   toProto(): api.IValue;
 }
@@ -246,7 +259,6 @@ export class Serializer {
   }
 }
 
-
 /**
  * Verifies that 'obj' is a plain JavaScript object that can be encoded as a
  * 'Map' in Firestore.
@@ -260,4 +272,101 @@ export function isPlainObject(input: unknown): input is object {
       isObject(input) &&
       (Object.getPrototypeOf(input) === Object.prototype ||
        Object.getPrototypeOf(input) === null));
+}
+
+/**
+ * Validates a JavaScript value for usage as a Firestore value.
+ *
+ * @private
+ * @param arg The argument name or argument index (for varargs methods).
+ * @param value JavaScript value to validate.
+ * @param desc A description of the expected type.
+ * @param path The field path to validate.
+ * @param options Validation options
+ * @param level The current depth of the traversal. This is used to decide
+ * whether deletes are allowed in conjunction with `allowDeletes: root`.
+ * @param inArray Whether we are inside an array.
+ * @throws when the object is invalid.
+ */
+export function validateUserInput(
+    arg: string|number, value: unknown, desc: string,
+    options: ValidationOptions, path?: FieldPath, level?: number,
+    inArray?: boolean): void {
+  if (path && path.size > MAX_DEPTH) {
+    throw new Error(
+        `${invalidArgumentMessage(arg, desc)} Input object is deeper than ${
+            MAX_DEPTH} levels or contains a cycle.`);
+  }
+
+  options = options || {};
+  level = level || 0;
+  inArray = inArray || false;
+
+  const fieldPathMessage = path ? ` (found in field ${path.toString()})` : '';
+
+  if (Array.isArray(value)) {
+    const arr = value as unknown[];
+    for (let i = 0; i < arr.length; ++i) {
+      validateUserInput(
+          arg, arr[i]!, desc, options,
+          path ? path.append(String(i)) : new FieldPath(String(i)), level + 1,
+          /* inArray= */ true);
+    }
+  } else if (isPlainObject(value)) {
+    const obj = value as object;
+    for (const prop in obj) {
+      if (obj.hasOwnProperty(prop)) {
+        validateUserInput(
+            arg, obj[prop]!, desc, options,
+            path ? path.append(new FieldPath(prop)) : new FieldPath(prop),
+            level + 1, inArray);
+      }
+    }
+  } else if (value === undefined) {
+    throw new Error(`${
+        invalidArgumentMessage(
+            arg, desc)} Cannot use "undefined" as a Firestore value${
+        fieldPathMessage}.`);
+  } else if (value instanceof DeleteTransform) {
+    if (inArray) {
+      throw new Error(`${invalidArgumentMessage(arg, desc)} ${
+          value.methodName}() cannot be used inside of an array${
+          fieldPathMessage}.`);
+    } else if (
+        (options.allowDeletes === 'root' && level !== 0) ||
+        options.allowDeletes === 'none') {
+      throw new Error(`${invalidArgumentMessage(arg, desc)} ${
+          value
+              .methodName}() must appear at the top-level and can only be used in update() or set() with {merge:true}${
+          fieldPathMessage}.`);
+    }
+  } else if (value instanceof FieldTransform) {
+    if (inArray) {
+      throw new Error(`${invalidArgumentMessage(arg, desc)} ${
+          value.methodName}() cannot be used inside of an array${
+          fieldPathMessage}.`);
+    } else if (!options.allowTransforms) {
+      throw new Error(`${invalidArgumentMessage(arg, desc)} ${
+          value.methodName}() can only be used in set(), create() or update()${
+          fieldPathMessage}.`);
+    }
+  } else if (value instanceof FieldPath) {
+    throw new Error(`${
+        invalidArgumentMessage(
+            arg,
+            desc)} Cannot use object of type "FieldPath" as a Firestore value${
+        fieldPathMessage}.`);
+  } else if (value instanceof DocumentReference) {
+    // Ok.
+  } else if (value instanceof GeoPoint) {
+    // Ok.
+  } else if (value instanceof Timestamp || value instanceof Date) {
+    // Ok.
+  } else if (value instanceof Buffer || value instanceof Uint8Array) {
+    // Ok.
+  } else if (value === null) {
+    // Ok.
+  } else if (typeof value === 'object') {
+    throw new Error(customObjectMessage(arg, value, path));
+  }
 }
